Migration Dashboard PrimeVue Confirm Dialog to UserDataTable for better request/error handling due to exception not chained from within dialog.
